FIR/UAST: commonize named/vararg expressions
